changement des repository Long en Integer

+++ server/src/main/java/fr/istic/sit/codisgroupea/repository/InterventionRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
  * SpringData repository for intervention entity.
  */
 @Repository
-public interface InterventionRepository extends JpaRepository<Intervention, Long> {
+public interface InterventionRepository extends JpaRepository<Intervention, Integer> {
 
     List<Intervention> findAllByOpened(boolean open);
 }
